ci: Build examples individually and fix warnings
Key changes:

- **CI Workflow (`.github/workflows/build-examples.yml`):**
  - Introduced a build matrix to compile each example individually.
  - Enabled stricter compiler warnings to improve code quality.
  - Corrected compiler invocation for MSVC to use `cl.exe`.

- **Code Refinements:**
  - `Core.hpp` & `Minimal.hpp`: Fixed a recursive call in `tryToPrintHelp()`.
  - `Get.hpp`: Added a `static_cast` to resolve a potential data loss warning during flag counting.
  - `GetTemplate.hpp` & `MinimalGet.hpp`: Updated the `tolower` transformation to safely handle character conversions.

+++ include/ArgLite/Get.hpp
@@ -34,7 +34,7 @@ inline unsigned Parser::countFlag_(
     auto longOptInfoArr  = getOptInfoArr(longOpt);
     auto shortOptInfoArr = getOptInfoArr(shortOpt);
 
-    return longOptInfoArr.size() + shortOptInfoArr.size();
+    return static_cast<unsigned>(longOptInfoArr.size() + shortOptInfoArr.size());
 }
 
 bool Parser::hasMutualExFlag_(HasMutualExArgs args, InternalData &data) {
